chore: extract `SpellListEntrySection` to a file
