+++ frontend/src/pages/character_sheet/panels/spells_list/RitualSpellsList.tsx
@@ -2,8 +2,7 @@ import BlurButton from "@common/BlurButton";
 import { Accordion, Badge, Box, Group, Stack, Text } from "@mantine/core";
 import { CastingSource, Character, Spell, SpellInnateEntry, SpellListEntry, SpellSlot } from "@typing/content";
 import { Dictionary } from "node_modules/cypress/types/lodash";
-import { SetterOrUpdater } from "recoil";
-import { SpellListEntrySection } from "../SpellsPanel";
+import SpellListEntrySection from "./SpellListEntrySection";
 
 export default function RitualSpellsList(props: {
   index: string;
