+++ frontend/src/pages/character_sheet/panels/spells_list/SpontaneousSpellsList.tsx
@@ -6,7 +6,8 @@ import { rankNumber } from "@utils/numbers";
 import { toLabel } from "@utils/strings";
 import { Dictionary } from "node_modules/cypress/types/lodash";
 import { SetterOrUpdater } from "recoil";
-import { SpellListEntrySection, SpellSlotSelect } from "../SpellsPanel";
+import { SpellSlotSelect } from "../SpellsPanel";
+import SpellListEntrySection from "./SpellListEntrySection";
 
 export default function SpontaneousSpellsList(props: {
   index: string;
